requirements.txt
PyYAML
dnacentersdk
pynetbox
fuzzywuzzy
levenshtein
python-slugify